snowflake_test/requirements.txt
snowflake-connector-python
pytest
